requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
distro==1.9.0
docutils==0.21.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.8
jiter==0.5.0
Kivy==2.3.0
kivy-deps.angle==0.4.0
kivy-deps.glew==0.3.1
kivy-deps.sdl2==0.7.0
Kivy-Garden==0.1.5
openai==1.42.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pypiwin32==223
python-dotenv==1.0.1
pywin32==306
requests==2.32.3
sniffio==1.3.1
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.2
